feat: add black formatter to CI
This change is needed to ensure that the codebase is formatted
consistently and follows the black formatter guidelines. This will help
improve code readability and maintainability.

This change adds a GitHub Actions workflow that runs the
black formatter on every push and pull request. The workflow uses the
rickstaa/action-black@v1 action to format the codebase and the
stefanzweifel/git-auto-commit-action@v5 action to automatically
commit the changes.

N/A